Setting up DB Context
[#84052834]
- Wired up the db context using the schema created by Josiah.
- Comment out the UsersContext as this likely wont be needed and was
adding junk to the web.config in the publish to Capstone.
- Setup the proper connection strings for the users portion and for the
entity framework portion.
- Removed some libraries that we will not be using

NOTE: The only two files you should ever have to touch with regards to
the db context (i.e. the folder DB_Data in the solution) is the edmx
file in the designer to pull in updates from the DB. DO NOT MESS WITH
THE .tt or .cs files since these are auto generated from the edmx.