Update Product table to have a FK to user
[#85184282]
Each product needs to be owned by a user, so I added a foreign key to
Copiosis-Setup.sql to add this key to the product table. Also updated
the edmx file to have this key.

Note: I will email a quick alter statement that EVERYONE needs to run on
their local machines