Rough work on Add Item Markup
[#83793980] Added some rough work on the add item creation, and added in
a library for handling the search/dropdown combo box since I think that
will be easier than creating our own. I think it will be easier to set
up the razor syntax once I have an idea of how the model it will be
using is set up, since I need to iterate through the data in order to
populate the search bar. Removed custom.js since anything that used it
is no longer in the solution