First pass at implementing AddItem
[#84265898]
Setup AddItem.cshtml to take a proper model to populate the drop down
menu. Also setup the post action to save a product away. Note: the view
still needs to pass along the product description and it needs some JS
to update the suggested gateway if the product template changes.