Wire up /EditItem and refactor some validation
[#84265964]
Set up the /EditItem endpoint so that it will return the item details to
the GET portion. From there the user can make changes to class, name,
description, and gateway. Then when they submit it will hit the POST
action with a model and the GUID of the item. Do some sanity checks to
make sure the model is valid and the item exists. Then update the item
in the database with the new values.