Custom Error Page started
[#85437358]
Added changes to Error.cshtml to display a custom error page.

Changed the appropriate line in <system.web> node of the Web.config file
to allow this routing.

Still need to make make the error messages more descriptive.

Note: in the _Layout.cshtml file, I was getting build errors, so I had
to wrap the script tags in the section blocks.