Stop transaction from being create if the gateway > NBR
[#86267276]
If you try to create a consumer transaction, and your NBR < the gateway
of the product. Throw an error saying you dont have enough NBR.
If you try to create a producer transaction, and the consumer you
selected has NBR < the gateway of your product. Throw an error saying
the consumer does not have enough NBR for this product.