[#86267002] Search by username
Hey Kyle, I made it so that it finds the user by their username now even
though we never display that data. This solution is kind of weird
though, take a look, and maybe revert it if you think it is too
convoluted.

Basically, I create another list that I populate at the same time as the
producers list, and since c# keeps lists ordered by the time they were
entered we can use that list and just match the index in order to get
the correct username out.

So I grab the index of the name from the select list and compare to the
username list in order to get the username we need.

Anyways take a look and let me know what you think