We shouldn't rely on the default item class being 11
[#85670460]
In most cases (i.e. 99%) the default item class will have id = 1.
However, adding this change will mean we dont have to rely on the id
being 1, instead look it up by name. I also fixed the Overview page, so
that the dropdown menu for default products is generated the same way
that normal products are.