Added success banner for EditClass view +
Used TempData dictionary to handle cases of edit/adding a class and
added a success panel banner after an admin edits a class.

I added an Equals() method to the AddClass model to be used to check if
the user presses the Submit button without changing any of the class'
properties. In this case, the admin is notified that no changes were
made to the particular class.