Added some styling to the navigation bar
The navigation bar items are now nav-pills and are set to 'active' when
the appropriate page is loaded. I attached id's on the nav items and I
am just adding 'active' classes to them for each page, rather than some
fancy JS function (which I could not get to work)

One note though is that when the window is really small, the nav-pills
collapsed aren't too pretty. It's a minor thing, though.