Fixed duplication of error message in AddItem
Account/AddItem validation errors will be caught before the form is
passed to the Controller.

The case where an item is both a Product AND Service type is an internal
error that will be caught in the Controller, after which an exception
will be thrown. Since these are radio buttons, this case should not
originate from this form.