Clean up and fixes for making a user an admin
[#87433932]
Simplified the idea, the user role will always exists on an account, it
is the admin role that can be added and removed. As such, users that are
currently admins can be demoted (i.e. lose the admin role). Users that
are not currently admins can be promoted (i.e. gain the admin role).
This can be controlled via buttons and some JS instead of the drop down
menus.

Also cleaned up the initial load for ViewUsers as this could be
simplified using the navigation properties on a role (i.e. role.users
will give you all of the users with that role). The last thing was to
only show distinct users (i.e. don't show kyle in the admin and
non-admin section)