Removed glyphs in exchange for responsive tooltips
I got frustrated with the glyphs span placement as it got awkward with
the labels. I added tooltips that are faster to display and nicer
looking when the mouse is hovered over the element.

You can see the changes in the Account/Overview page in the 'Create
Transaction' button group and in the Account/Manage page.

I am proposing we move towards this style of tooltips @kylegalbraith