Dropdown should be working again
[Delivers #88245220] It works again, and also I cleared out some
unnecessary stuff from when we didn't have usernames. There is still
some stuff we can trim probably between the fetchProducerItems and the
ajax call. The fetchInitialProducerItems function may no longer be
needed either. I'm just leaving it as is for now though, and I'll play
around with it some more if I have time.